Project-dagadaga/blockchain/requirements.txt
# Blockchain & Payment Service Requirements

# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2

# Blockchain dependencies
pycardano==0.11.0
blockfrost-python==0.5.4

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Security
cryptography==41.0.8
python-jose[cryptography]==3.3.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0